Refactor codegen script into an abstract final class
